pyproject.toml
[project]
name = "marketing-agents"
version = "0.1.0"
description = "A swarm of research-focused marketing agents with sequential pipeline execution"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tavily>=0.2.0",
    "httpx>=0.27.0",
    "tiktoken>=0.7.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["marketing_agents*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python-version = "3.9"
strict = false
